requirements.txt
# base.txt is a pip requirements file which describes the necessary
# distributions required to run this project. If you need something that is
# project specific not listed here use project.txt. You can, of course, update
# versions of distributions here if needed.

--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/
--extra-index-url=http://dist.pinaxproject.com/fresh-start/

Django==1.4.3
django-user-accounts==1.0b7
symposion==1.0b1.dev45
pinax-theme-bootstrap==2.2.2
pinax-theme-bootstrap-account==1.0b2
django-forms-bootstrap==2.0.3.post1
metron==0.1 # 0.2.dev3
pinax-utils==1.0b1.dev3
django-debug-toolbar==0.9.4
django-mailer==0.2a1
django-timezones==0.2
pytz==2012g
django-model-utils==1.1.0
django-modeltranslation==0.6.1

django-taggit==0.9.3
django-reversion==1.6.5
django-markitup==1.0.0
markdown==2.2.0
django-sitetree==0.9.1
PIL==1.1.7
easy-thumbnails==1.1
